@@ -3,6 +3,7 @@ plugins {
     id 'kotlin-android'
     id 'kotlin-kapt'
     id 'kotlin-parcelize'
+    id 'org.jetbrains.kotlin.android'
 }
 
 android {
@@ -37,54 +38,71 @@ android {
     }
 
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_19
+        targetCompatibility JavaVersion.VERSION_19
     }
+
     kotlinOptions {
-        jvmTarget = '1.8'
+        jvmTarget = '19'
+    }
+
+    testOptions {
+        unitTests.all {
+            useJUnitPlatform()
+        }
     }
 }
 
 dependencies {
+    implementation 'androidx.core:core-ktx:+'
+
     // Lib versions
     def room_version = "2.4.3"
     def lifecycle_version = "2.5.1"
     def test_version = "1.2.0"
+    def faker_version = "1.13.0"
     def arch_version = "1.0.0-alpha5"
 
     // Annotation processor
     kapt "androidx.room:room-compiler:$room_version"
 
     // Core
-    implementation 'androidx.core:core-ktx:1.9.0'
+    implementation "androidx.core:core-ktx:1.9.0"
+    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation "androidx.lifecycle:lifecycle-runtime-ktx:$lifecycle_version"
 
     // UI
-    implementation 'androidx.appcompat:appcompat:1.4.1'
-    implementation 'com.google.android.material:material:1.6.0'
+    implementation "androidx.appcompat:appcompat:1.4.1"
+    implementation "com.google.android.material:material:1.6.0"
     implementation "androidx.recyclerview:recyclerview:1.2.1"
-    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
-    implementation 'com.android.databinding:viewbinding:7.3.1'
+    implementation "androidx.constraintlayout:constraintlayout:2.1.4"
+    implementation "com.android.databinding:viewbinding:7.3.1"
     implementation 'io.coil-kt:coil:2.2.2'
 
     // Room
     implementation "androidx.room:room-runtime:$room_version"
     implementation "androidx.room:room-ktx:$room_version"
+    androidTestImplementation "androidx.room:room-testing:$room_version"
+    androidTestImplementation "android.arch.persistence.room:testing:$arch_version"
 
     // Coroutines
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.9'
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.9"
+    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.6.4"
 
     // DataStore (Preferences)
     implementation "androidx.datastore:datastore-preferences:1.0.0"
 
     // Tests
-    implementation 'androidx.test.ext:junit-ktx:1.1.5'
-    testImplementation 'junit:junit:4.+'
-    testImplementation 'org.amshove.kluent:kluent-android:1.72'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.4'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.0'
+    implementation "io.github.serpro69:kotlin-faker:$faker_version"
+    testImplementation "org.amshove.kluent:kluent-android:1.72"
+    testImplementation "io.mockk:mockk:1.13.4"
+
+    // JUnit
+    testImplementation "org.junit.jupiter:junit-jupiter-api:5.9.2"
+    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.9.2'
+
+    androidTestImplementation "androidx.test.ext:junit:1.1.5"
+    androidTestImplementation "androidx.test.espresso:espresso-core:3.5.0"
     androidTestImplementation "androidx.test:runner:$test_version"
     androidTestImplementation "androidx.test:rules:$test_version"
-    androidTestImplementation "androidx.room:room-testing:$room_version"
-    androidTestImplementation "android.arch.persistence.room:testing:$arch_version"
 }
