@@ -1,7 +1,7 @@
 package com.example.orgs.model
 
 import org.amshove.kluent.*
-import org.junit.Test
+import org.junit.jupiter.api.Test
 import java.math.BigDecimal
 
 class ProdutoTest {
